Separate test binary from main exec
- Create a test_optimism_preimage_read binary with its own compilation path
- Create a common library to do CLI parsing to be used in both executables
- Update test script accordingly (with additional small fixes)
