+++ optimism/Cargo.toml
@@ -12,6 +12,10 @@ license = "Apache-2.0"
 [lib]
 path = "src/lib.rs"
 
+[[bin]]
+name = "test_optimism_preimage_read"
+path = "src/test_preimage_read.rs"
+
 [dependencies]
 ark-bn254 = { version = "0.3.0" }
 kimchi = { path = "../kimchi", version = "0.1.0", features = [ "bn254" ] }
