+++ optimism/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod cannon;
+pub mod cannon_cli;
 pub mod mips;
 pub mod preimage_oracle;
