+++ optimism/test_preimage_read.sh
@@ -2,7 +2,7 @@
 # Test preimage read from VM request to server response
 # Usage: test_preimage_read.sh [OP_DB_DIR] [NETWORK_NAME]
 
-source rpcs.sh 
+source rpcs.sh
 
 set +u
 if [ -z "${FILENAME}" ]; then
@@ -12,11 +12,10 @@ set -u
 
 source $FILENAME
 
-
 set -x
 
 # Location of prepopulated OP DB
-# OP_PROGRAM_DATA_DIR is generated by ./generate-config.sh 
+# OP_PROGRAM_DATA_DIR is generated by ./generate-config.sh
 # It makes sense to use it as default value here
 DATADIR=${1:-${OP_PROGRAM_DATA_DIR}}
 NETWORK=${2:-sepolia}
@@ -36,10 +35,10 @@ NETWORK=${2:-sepolia}
 #     --l2.blocknumber ${L2_BLOCK_NUMBER}       #
 #################################################
 
-
-
-RUST_LOG=debug cargo run -r --bin kimchi_optimism -- --preimage-db-dir ${DATADIR}  -- \
-    ${OP_DIR}/op-program/bin/op-program \
+# Run test with debug on
+RUST_LOG=debug cargo run -r --bin test_optimism_preimage_read -- \
+    --preimage-db-dir ${DATADIR} -- \
+    op-program \
     --log.level DEBUG \
     --l1 ${L1RPC} \
     --l2 ${L2RPC} \
